Cleanup product related scripts - WIP
