--- scripts/v2-migration/output/sunsetProducts.json
@@ -1,1 +1,0 @@
-["0xF5DCe57282A584D2746FaF1593d3121Fcac444dC","0x8B3d70d628Ebd30D4A2ea82DB95bA2e906c71633","0x5504a1d88005236147EC86C62cfb53043bD1276a","0x4Ddc2D193948926D02f9B1fE9e1daa0718270ED5","0x080bf510FCbF18b91105470639e9561022937712","0x16de59092dAE5CcF4A1E6439D611fd0653f0Bd01","0xAF350211414C5DC176421Ea05423F0cC494261fB","0x5d3a536E4D6DbD6114cc1Ead35777bAB948E3643","0x2157A7894439191e520825fe9399aB8655E0f708","0x9f8F72aA9304c8B593d555F12eF6589cC3A579A2","0x6e95C8E8557AbC08b46F3c347bA06F8dC012763f","0xb1CD6e4153B2a390Cf00A6556b0fC1458C4A5533","0x29fe7D60DdF151E5b52e5FAB4f1325da6b2bD958","0x4a57E687b9126435a9B19E4A802113e266AdeBde","0x519b70055af55A007110B4Ff99b0eA33071c720a","0x802275979B020F0ec871c5eC1db6e412b72fF20b","0xb7896fce748396EcFC240F5a0d3Cc92ca42D7d84","0x932773aE4B661029704e731722CF8129e1B32494","0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39","0x2C4Bd064b998838076fa341A83d007FC2FA50957","0x6B175474E89094C44Da98b954EedeAC495271d0F","0xD5D2b9e9bcd172D5fC8521AFd2C98Dd239F5b607","0x241e82C79452F51fbfc89Fac6d912e021dB1a3B7","0x12D66f87A04A9E220743712cE6d9bB1B5616B8Fc","0x5d22045DAcEAB03B158031eCB7D9d06Fad24609b","0x498b3BfaBE9F73db90D252bCD4Fa9548Cd0Fd981","0x448a5065aeBB8E423F0896E6c5D525C040f59af3","0xe80d347DF1209a76DD9d2319d62912ba98C54DDD","0xB27F1DB0a7e473304A5a06E54bdf035F671400C0","0x1E0447b19BB6EcFdAe1e4AE1694b0C3659614e4e","0xc0a47dFe034B400B47bDaD5FecDa2621de6c4d95","0x72338b82800400F5488eCa2B5A37270ba3B7A111","0xc1D2819CE78f3E15Ee69c6738eB1B400A26e632A","0x10eC0D497824e342bCB0EDcE00959142aAa766dD","0x3dfd23A6c5E8BbcFc9581d2E864a68feb6a076d3","0x45F783CCE6B7FF23B2ab2D70e416cdb7D6055f51","0x11111254369792b2Ca5d084aB5eEA397cA8fa48B","0xF92C1ad75005E6436B4EE84e88cB23Ed8A290988","0xb529964F86fbf99a6aA67f72a27e59fA3fa4FEaC","0x77208a6000691E440026bEd1b178EF4661D37426","0x12f208476F64De6e6f933E55069Ba9596D818e08","0x78751B12Da02728F467A44eAc40F5cbc16Bd7934","0x7fC77b5c7614E1533320Ea6DDc2Eb61fa00A9714","0x9424B1412450D0f8Fc2255FAf6046b98213B76Bd","0xD36132E0c1141B26E62733e018f12Eb38A7b7678","0x86969d29F5fd327E1009bA66072BE22DB6017cC6","0x5f9AE054C7F0489888B1ea46824b4B9618f8A711","0x1fd169A4f5c59ACf79d0Fd5d91D1201EF1Bce9f1","0xAFcE80b19A8cE13DEc0739a1aaB7A028d6845Eb3","0x2a0c0DBEcC7E4D658f48E01e3fA353F44050c208","0x9AAb3f75489902f3a48495025729a0AF77d4b11e","0x3e532e6222afe9Bcf02DCB87216802c75D5113aE","0x02285AcaafEB533e03A7306C55EC031297df9224","0x0e2298E3B3390e3b945a5456fBf59eCc3f55DA16","0x3fE7940616e5Bc47b0775a0dccf6237893353bB4","0x71CD6666064C3A1354a3B4dca5fA1E2D3ee7D303","0xe20A5C79b39bC8C363f0f49ADcFa82C2a01ab64a","0xe9778E69a961e64d3cdBB34CF6778281d34667c2","0x4C39b37f5F20a0695BFDC59cf10bd85a6c4B7c30","0x3A97247DF274a17C59A3bd12735ea3FcDFb49950","0x26aaD4D82f6c9FA6E34D8c1067429C986A055872","0xCB876f60399897db24058b2d58D0B9f713175eeF","0xa4c8d221d8BB851f83aadd0223a8900A6921A349","0xB94199866Fe06B535d019C11247D3f921460b91A","0x878F15ffC8b894A1BA7647c7176E4C01f74e140b","0xfA5047c9c78B8877af97BDcb85Db743fD7313d4a","0x3d5BC3c8d13dcB8bF317092d84783c2697AE9258","0x7a9701453249e84fd0D5AfE5951e9cBe9ed2E90f","0x67B66C99D3Eb37Fa76Aa3Ed1ff33E8e39F0b9c7A","0xc2EdaD668740f1aA35E4D8f227fB8E17dcA888Cd","0xedfC81Bf63527337cD2193925f9C0cF2D537AccA","0xA51156F3F1e39d1036Ca4ba4974107A1C1815d1e","0x9abd0b8868546105F6F48298eaDC1D9c82f7f683","0xE75D77B1865Ae93c7eaa3040B038D7aA7BC02F70","0x7C06792Af1632E77cb27a558Dc0885338F4Bdf8E","0xa4F1671d3Aee73C05b552d57f2d16d3cfcBd0217","0x5D9972dD3Ba5602574ABeA6bF9E1713568D49903","0x284D7200a0Dabb05ee6De698da10d00df164f61d","0x4B8d90D68F26DEF303Dcb6CFc9b63A1aAEC15840","0xc21D353FF4ee73C572425697f4F5aaD2109fe35b","0x99c666810bA4Bf9a4C2318CE60Cb2c279Ee2cF56","0xA625AB01B08ce023B2a342Dbb12a16f2C8489A8F","0x08FB62c84909dA3Aa5F59E01763E5FDC62De76e9","0xc57D000000000000000000000000000000000001","0xc57d000000000000000000000000000000000004","0xefa94DE7a4656D787667C749f7E1223D71E9FD88","0x0CED6166873038Ac0cc688e7E6d19E2cBE251Bf0","0x0000000000000000000000000000000000000002","0x0000000000000000000000000000000000000003","0x0000000000000000000000000000000000000005","0x0000000000000000000000000000000000000006","0x0000000000000000000000000000000000000007","0x0000000000000000000000000000000000000008","0x0000000000000000000000000000000000000009","0x0000000000000000000000000000000000000010","0x0000000000000000000000000000000000000011","0x0000000000000000000000000000000000000012","0x0000000000000000000000000000000000000013","0x0000000000000000000000000000000000000014","0x0000000000000000000000000000000000000015","0xaE7b92C8B14E7bdB523408aE0A6fFbf3f589adD9","0xd89a09084555a7D0ABe7B111b1f78DFEdDd638Be","0x2BB8de958134AFd7543d4063CaFAD0b7c6de08BC","0x66357dCaCe80431aee0A7507e2E361B7e2402370","0x0000000000000000000000000000000000000017","0x0000000000000000000000000000000000000024"]
