+++ scripts/v2-migration/products/input/covers.json
