+++ scripts/v2-migration/products/output/v2ProductAddresses.json
@@ -67,6 +67,7 @@
   "0x0000000000000000000000000000000000000032",
   "0x1F573D6Fb3F13d689FF844B4cE37794d79a7FF1C",
   "0x00000000219ab540356cBB839Cbe05303d7705Fa",
+  "0xc57D000000000000000000000000000000000001",
   "0xC57D000000000000000000000000000000000002",
   "0xC57d000000000000000000000000000000000003",
   "0xC57D000000000000000000000000000000000005",
