+++ scripts/version-data/create-mainnet-version-data.js
