+++ scripts/version-data/create-testnet-version-data.js
