+++ scripts/version-data/version-data-template.js
